+++ TelegramBotAPI-core/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/CommonSupergroupEventMessage.kt
@@ -1,14 +1,19 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.User
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.SupergroupChat
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.ChatEvents.abstracts.SupergroupEvent
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.ChatEventMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.SupergroupEventMessage
 import com.soywiz.klock.DateTime
 
-data class SupergroupEventMessage(
+@Deprecated("Renamed", ReplaceWith("CommonSupergroupEventMessage"))
+typealias SupergroupEventMessage = CommonSupergroupEventMessage
+
+data class CommonSupergroupEventMessage(
     override val messageId: MessageIdentifier,
+    override val user: User,
     override val chat: SupergroupChat,
     override val chatEvent: SupergroupEvent,
     override val date: DateTime
-) : ChatEventMessage
+) : SupergroupEventMessage