+++ TelegramBotAPI-core/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/GroupEventMessage.kt
@@ -1,0 +1,3 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
+
+interface GroupEventMessage : ChatEventMessage, FromUserMessage